requirements.txt
cycler==0.11.0
fonttools==4.28.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
opencv-python==4.5.4.58
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2021.3
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
typing-extensions==4.0.0
